pyproject.toml
[tool.poetry]
name = "flask-app-with-docker"
version = "1.0"
description = "This is a flask app dockerized, for training purposes"
authors = ["afro-juliano <afro.juliano@gmail.com>"]
license = "GNU General Public License v3.0"
readme = "README.md"
packages = [{include = "flask_app_with_docker"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
black = "^23.7.0"
flake8 = "^6.1.0"
flask-mysql = "^1.5.2"
db = "^0.1.1"
mysql-connector-python = "^8.1.0"
mysql-connector = "^2.2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
